% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dems.R
\name{chm_from_coregistered_dsm_dtm}
\alias{chm_from_coregistered_dsm_dtm}
\title{Calculate Canopy Height Model (CHM) from a coregistered DSM and DTM}
\usage{
chm_from_coregistered_dsm_dtm(dsm, dtm, res = 0.12, bounds = NULL)
}
\arguments{
\item{dsm}{A terra SpatRaster representing the Digital Surface Model.}

\item{dtm}{A terra SpatRaster representing the Digital Terrain Model.}

\item{res}{The desired resolution of the CHM. Default is 0.12 m.}

\item{bounds}{An sf object representing the bounds to crop the CHM to.}
}
\value{
A terra SpatRaster representing the Canopy Height Model.
}
\description{
This function calculates the Canopy Height Model (CHM) by subtracting the Digital Terrain Model
(DTM) from the Digital Surface Model (DSM).
}
\examples{
chm_from_coregistered_dsm_dtm(dsm_raster, dtm_raster)
}
