% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter-tuning.R
\name{create_parameter_grid}
\alias{create_parameter_grid}
\title{Create a grid search of parameters and optionally take a random subset}
\usage{
create_parameter_grid(parameter_list, n_random_samples = NULL)
}
\arguments{
\item{parameter_list}{A named list of parameters. Each element is either a list of values or a single value}

\item{n_random_samples}{How many samples to draw from the cross product list. If NULL, all elements will be taken}
}
\value{
A dataframe where each row contains one value from each parameter drawn from the list
}
\description{
Create a grid search of parameters and optionally take a random subset
}
