% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual-tree-detection.R
\name{detect_trees2}
\alias{detect_trees2}
\title{Delineate individual trees from canopy height model raster}
\usage{
detect_trees2(
  chm,
  ws,
  algorithm = lidR::lmf(ws = ws),
  uniqueness = "incremental",
  min_ht = 5
)
}
\arguments{
\item{chm}{A terra SpatRaster representing the canopy height model.}

\item{ws}{moving window size. This indicates as number of pixels to consider.}
}
\value{
A spatial point DataFrame representing the tree locations, height, and tree ID.
}
\description{
This function uses pixel based height parameters and their distribution
around a given neighborhood to find individual tree locations and later the tree crown area and tree heights.
}
