% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-detection-accuracy-assessment.R
\name{eval_res_dist}
\alias{eval_res_dist}
\title{Return the euclidian distance between the predicted and true shift}
\usage{
eval_res_dist(predicted_shift, map_config)
}
\arguments{
\item{predicted_shift}{The predicted shift with shift_x and shift_y attributes}

\item{map_config}{The true shifts with shift_x and shift_y attributes}
}
\value{
A scalar distance
}
\description{
Return the euclidian distance between the predicted and true shift
}
