% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-imagery-metadata_emp.R
\name{extract_altitude_asl}
\alias{extract_altitude_asl}
\title{Returns altitude above sea level (asl) in meters}
\usage{
extract_altitude_asl(exif)
}
\arguments{
\item{exif}{the exif metadata file}

\item{altitude_asl}{altitude above sea level in meters}
}
\value{
altitude above sea level in meters
}
\description{
Returns altitude above sea level (asl) in meters to include in image-level metadata collation.
}
\examples{
extract_altitude_asl(exif)

}
