% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata-extraction_imagery_dataset.R
\name{extract_area_and_density}
\alias{extract_area_and_density}
\title{Extracts the the area of the mission footprint and the image density of the dataset}
\usage{
extract_area_and_density(exif, mission_polygon)
}
\arguments{
\item{exif}{the exif metadata file}

\item{mission_polygon}{the sf polygon object of the mission footprint}
}
\value{
the area of the mission footprint (ha) and the image density based on image count and footprint (img/ha)
}
\description{
area_ha: Area of mission footprint. Units: ha. image_density: the image density computed with image count and footprint. Units: img/ha
}
\examples{
extract_area_and_density(exif)

}
