% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-dataset-metadata_emp.R
\name{extract_area_ha}
\alias{extract_area_ha}
\title{Extracts the area of the mission footprint}
\usage{
extract_area_ha(exif, image_merge_distance)
}
\arguments{
\item{exif}{the exif metadata file}
}
\value{
the area of mission footprint. Units: ha.
}
\description{
Area of mission footprint. Units: ha.
}
\examples{
extract_area_ha(exif)

}
