% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata-extraction_imagery_dataset.R
\name{extract_area_ha_and_image_density}
\alias{extract_area_ha_and_image_density}
\title{Extracts the the area of the mission footprint and the image density of the dataset}
\usage{
extract_area_ha_and_image_density(exif, image_merge_distance)
}
\arguments{
\item{exif}{the exif metadata file}

\item{image_merge_distance}{the horizontal distance between images below which they are merged into one mission polygon}
}
\value{
the area of the mission footprint (ha) and the image density based on image count and footprint (img/ha)
}
\description{
area_ha: Area of mission footprint. Units: ha. image_density: the image density computed with image count and footprint. Units: img/ha
}
\examples{
extract_area_ha_and_image_density(exif)

}
