% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-imagery-metadata_emp.R
\name{extract_dataset_id}
\alias{extract_dataset_id}
\title{Extract dataset id}
\usage{
extract_dataset_id(exif)
}
\arguments{
\item{dataset_id}{dataset id}
}
\value{
dataset id
}
\description{
Pulls the dataset id to include in image-level metadata collation
}
\examples{
extract_dataset_id(exif)

}
