% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-imagery-metadata_emp.R
\name{extract_dataset_id_image_level}
\alias{extract_dataset_id_image_level}
\title{Extract dataset id}
\usage{
extract_dataset_id_image_level(exif)
}
\arguments{
\item{exif}{the exif metadata file}
}
\value{
dataset id
}
\description{
Pulls the dataset id to include in image-level metadata collation
}
\examples{
extract_dataset_id(exif)

}
