% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata-extraction_imagery_perimage.R
\name{extract_dataset_id_perimage}
\alias{extract_dataset_id_perimage}
\title{Extract dataset id}
\usage{
extract_dataset_id_perimage(exif)
}
\arguments{
\item{exif}{the exif metadata file}
}
\value{
dataset id
}
\description{
Pulls the dataset id, to include in image-level metadata collation
}
\examples{
extract_dataset_id(exif)

}
