% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata-extraction_imagery_dataset.R
\name{extract_dataset_id_summary}
\alias{extract_dataset_id_summary}
\title{Extract dataset id}
\usage{
extract_dataset_id_summary(metadata)
}
\arguments{
\item{metadata}{the metadata metadata file}
}
\value{
dataset id
}
\description{
Pulls the dataset id to include in dataset-level metadata. Relies on dataset_id being set
properly by the prep_metadata function.
}
\examples{
extract_dataset_id(metadata)

}
