% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-dataset-metadata_emp.R
\name{extract_dataset_metadata_emp}
\alias{extract_dataset_metadata_emp}
\title{Returns a collection of dataset-level metadata parameters}
\usage{
extract_dataset_metadata_emp(exif_filepath, image_merge_distance)
}
\arguments{
\item{exif_file}{the exif filepath (before being prepared to pass to the functions in the wrapper)

#' @param image_merge_distance The horizontal distance between images below which they are merged into one mission polygon}
}
\value{
a data.frame of dataset_id_dataset_level, image_count, file_size, percent_images_rtk, white_balance_mode_derived, white_balance_mode_prop_derived, exposure_median_derived, exposure_stdev_derived, image_density, and area_ha
}
\description{
Returns a collection of dataset-level metadata parameters
}
\examples{
extract_dataset_metadata_emp(exif)

}
