% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-imagery-metadata_emp.R
\name{extract_datatime_local}
\alias{extract_datatime_local}
\title{Extract local date and time of image collection}
\usage{
extract_datatime_local(exif)
}
\arguments{
\item{exif}{the exif metadata file}
}
\value{
local date and time of image collection
}
\description{
Pulls the local date and time of image collection (Format: YYYYMMDD HHMMSS) to include in image-level metadata collation
}
\examples{
extract_datatime_local(exif)

}
