% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata-extraction_imagery_perimage.R
\name{extract_datetime_local}
\alias{extract_datetime_local}
\title{Extract local date and time of image collection}
\usage{
extract_datetime_local(
  exif,
  candidate_columns = c("capture_datetime", "GPSDateTime")
)
}
\arguments{
\item{exif}{the exif metadata file}
}
\value{
local date and time of image collection
}
\description{
Pulls the local date and time of image collection (Format: YYYYMMDD HHMMSS) to include in image-level metadata collation
}
\examples{
extract_datatime_local(exif)

}
