% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagery-metadata-extraction_emp.R,
%   R/metadata-extraction_imagery_dataset.R
\name{extract_exposure}
\alias{extract_exposure}
\title{Returns exposure (units: seconds)}
\usage{
extract_exposure(exif)

extract_exposure(exif)
}
\arguments{
\item{exif}{the exif metadata file}
}
\value{
exposure (units: seconds)

the median exposure time across all images in the dataset. Units: sec
}
\description{
Returns exposure time in seconds to include in image-level metadata collation.

The median exposure time across all images in the dataset. Units: sec
}
\examples{
extract_exposure(exif)

extract_exposure(exif)

}
