% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata-extraction_imagery_dataset.R
\name{extract_exposure_median_derived}
\alias{extract_exposure_median_derived}
\title{Extracts the median exposure time}
\usage{
extract_exposure_median_derived(exif)
}
\arguments{
\item{exif}{the exif metadata file}
}
\value{
the median exposure time across all images in the dataset. Units: sec
}
\description{
The median exposure time across all images in the dataset. Units: sec
}
\examples{
extract_exposure_median_derived(exif)

}
