% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-dataset-metadata_emp.R
\name{extract_exposure_stdev_derived}
\alias{extract_exposure_stdev_derived}
\title{Extracts the standard deviation of exposure time}
\usage{
extract_exposure_stdev_derived(exif)
}
\arguments{
\item{exif}{the exif metadata file}
}
\value{
the standard deviation of exposure time across all images in the dataset. Units: sec
}
\description{
The standard deviation of exposure time across all images in the dataset. Units: sec
}
\examples{
extract_exposure_stdev_derived(exif)

}
