% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata-extraction_imagery_dataset.R
\name{extract_file_size_summary}
\alias{extract_file_size_summary}
\title{Extracts total file size}
\usage{
extract_file_size_summary(metadata)
}
\arguments{
\item{metadata}{the metadata metadata file}
}
\value{
total file size
}
\description{
Total file size of all images in set. Units: GB.
}
\examples{
extract_file_size(metadata)

}
