% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata-extraction_imagery_dataset.R
\name{extract_image_count}
\alias{extract_image_count}
\title{Extract number of images in an metadata}
\usage{
extract_image_count(metadata)
}
\arguments{
\item{metadata}{the metadata metadata file}
}
\value{
image count
}
\description{
Pulls the image count to include in dataset-level metadata
}
\examples{
extract_image_count(metadata)

}
