% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-dataset-metadata_emp.R
\name{extract_image_density}
\alias{extract_image_density}
\title{Extracts the image density of the dataset}
\usage{
extract_image_density(exif, image_merge_distance)
}
\arguments{
\item{exif}{the exif metadata file}
}
\value{
the image density based on image count and footprint. Units: img/ha
}
\description{
The image density computed with image count and footprint. Units: img/ha
}
\examples{
extract_image_density(exif)

}
