% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata-extraction_imagery_perimage.R
\name{extract_imagery_perimage_metadata}
\alias{extract_imagery_perimage_metadata}
\title{Extract image-level metadata parameters from EXIF datafram}
\usage{
extract_imagery_perimage_metadata(exif, platform_name, plot_flightpath = FALSE)
}
\arguments{
\item{exif_file}{the exif filepath (before being prepared to pass to the functions in the
wrapper)}

\item{input_type}{the type of input, either "dataframe" or "filepath" to a .csv file}
}
\value{
a data.frame of dataset_id_image_level, datetime_local, lat, lon, rtk_fix, accuracy_x, accuracy_y, camera_pitch, camera_roll, camera_yaw, exposure, aperture, iso, white_balance, received_image_path, and altitude_asl
}
\description{
Extract image-level metadata parameters from EXIF datafram
}
\examples{
extract_metadata_emp(exif)

}
