% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata-extraction_imagery_perimage.R
\name{extract_iso}
\alias{extract_iso}
\title{Returns ISO}
\usage{
extract_iso(exif)
}
\arguments{
\item{exif}{the exif metadata file}
}
\value{
ISO
}
\description{
Returns ISO to include in image-level metadata collation.
}
\examples{
extract_iso(exif)

}
