% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata-extraction_imagery_perimage.R
\name{extract_lon_lat}
\alias{extract_lon_lat}
\title{Extract latitude and longitude of image collection location}
\usage{
extract_lon_lat(exif)
}
\arguments{
\item{exif}{the exif metadata file}
}
\value{
latitude and longitude of image collection location
}
\description{
Pulls the latitude and longitude of image collection location (Format: dd.dddddddd) to include in image-level metadata collation
}
\examples{
extract_lat_lon(exif)

}
