% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-imagery-metadata_emp.R
\name{extract_metadata_emp}
\alias{extract_metadata_emp}
\title{Returns a collection of image-level metadata parameters}
\usage{
extract_metadata_emp(exif_file)
}
\arguments{
\item{exif_file}{the exif filepath (before being prepared to pass to the functions in the wrapper)}
}
\value{
a data.frame of dataset_id, datatime_local, lat, lon, rtk_fix, accuracy_x, accuracy_y, camera_pitch, camera_roll, camera_yaw, exposure, aperture, iso, white_balance, received_image_path, and altitude_asl
}
\description{
Returns a collection of image-level metadata parameters
}
\examples{
extract_metadata_emp(exif)

}
