% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata-extraction_imagery_dataset.R
\name{extract_percent_images_rtk}
\alias{extract_percent_images_rtk}
\title{Extracts percent of images with an RTK fix}
\usage{
extract_percent_images_rtk(exif)
}
\arguments{
\item{exif}{the exif metadata file}
}
\value{
percent of images with an RTK fix
}
\description{
Percent of images with an RTK fix
}
\examples{
extract_precent_images_rtk(exif)

}
