% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-imagery-metadata_emp.R
\name{extract_pitch_roll_yaw}
\alias{extract_pitch_roll_yaw}
\title{Returns camera pitch, camera roll, and camera yaw. Units: degrees.}
\usage{
extract_pitch_roll_yaw(exif)
}
\arguments{
\item{camera_pitch}{returns the camera pitch in degrees (degrees up from nadir)}

\item{camera_roll}{returns the camera roll in degrees (degrees clockwise from up)}

\item{camera_yaw}{returns the camera yaw in degrees (degrees right from true north)}

\item{pitch_roll_yaw}{returns a data.frame of the camera_pitch, camera_roll, and camera_yaw values for each image}
}
\value{
camera pitch, camera roll, and camera yaw. Units: degrees.
}
\description{
Returns camera_pitch (Units: deg, degrees up from nadir), camera_roll (Units: deg, degrees clockwise from up), and camera_yaw (Units: deg, degrees right from true north) to include in image-level metadata collation.
}
\examples{
extract_pitch_roll_yaw(exif)

}
