% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-imagery-metadata_emp.R
\name{extract_received_image_path}
\alias{extract_received_image_path}
\title{Returns image path}
\usage{
extract_received_image_path(exif)
}
\arguments{
\item{exif}{the exif metadata file}

\item{received_image_path}{the received image path}
}
\value{
received image path
}
\description{
Returns path in as-received dataset, with the top level being the folder named with the dataset ID, to include in image-level metadata collation.
}
\examples{
extract_received_image_path(exif)

}
