% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagery-metadata-extraction_emp.R
\name{extract_white_balance}
\alias{extract_white_balance}
\title{Returns white balance, either auto or manual}
\usage{
extract_white_balance(exif)
}
\arguments{
\item{exif}{the exif metadata file}
}
\value{
white balance
}
\description{
Returns white balance (auto vs. manual) to include in image-level metadata collation.
}
\examples{
extract_white_balance(exif)

}
