% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-dataset-metadata_emp.R
\name{extract_white_balance_mode_derived}
\alias{extract_white_balance_mode_derived}
\title{Extracts the most common white balance setting}
\usage{
extract_white_balance_mode_derived(exif)
}
\arguments{
\item{exif}{the exif metadata file}
}
\value{
the most common white balance setting
}
\description{
The most common white balance setting across all images in the project. Options: auto and manual
}
\examples{
extract_white_balance_mode_derived(exif)

}
