% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-dataset-metadata_emp.R
\name{extract_white_balance_mode_prop_derived}
\alias{extract_white_balance_mode_prop_derived}
\title{Extracts the proportion of images with a white balance setting that matches the most common white balance setting in the dataset}
\usage{
extract_white_balance_mode_prop_derived(exif)
}
\arguments{
\item{exif}{the exif metadata file}
}
\value{
the proportion of images matching the most common white balance setting
}
\description{
The proportion of images with a white balance setting that matches the modal white balance setting in the dataset
}
\examples{
extract_white_balance_mode_prop_derived(exif)

}
