% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations_emp.R
\name{flightpath_visualization_emp}
\alias{flightpath_visualization_emp}
\title{Returns an interactive map that shows the drone flightpath with locations of images. Altitude above sea level and collection time color schemes can be toggled on and off.}
\usage{
flightpath_visualization_emp(exif_file)
}
\arguments{
\item{exif_file}{the exif filepath (before being prepared to pass to the functions in the wrapper)}
}
\value{
an interactive map that shows the drone flightpath with locations of images. Altitude above sea level and collection time color schemes can be toggled on and off.
}
\description{
Returns an interactive map that shows the drone flightpath with locations of images. Altitude above sea level and collection time color schemes can be toggled on and off.
}
\examples{
flightpath_visualization_emp(exif_file)

}
