% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter-tuning.R
\name{test_tree_map_alignment}
\alias{test_tree_map_alignment}
\title{Test the different tree alignment algorithms in a variety of situations}
\usage{
test_tree_map_alignment(
  map_params,
  alignment_methods,
  per_method_alignment_arguments = NULL,
  alignment_method_names = NULL,
  n_random_samples = NULL
)
}
\arguments{
\item{map_params}{Named list. The names correspond to arguments to simulate_tree_maps. The values
correspond to potential values of this argument to be tried}

\item{alignment_methods}{List of alignment algorithms}

\item{per_method_alignment_arguments}{A list of named lists, one for each alignment method.
The arguments in each list will be passed to the corresponding function}

\item{alignment_method_names}{The human-readable names for each method}

\item{n_random_samples}{The number of random samples to draw from the grid of paramters}
}
\value{
Named list with the following fields
per_method_results: Named list, where each name is from alignment_method_names. The value is a data
frame of results from the experiments using that method.
all_map_param_configurations: Dataframe of paramters used in experiments, one row per trial
}
\description{
Test the different tree alignment algorithms in a variety of situations
}
