% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-map-alignment.R
\name{transform_points_by_homography}
\alias{transform_points_by_homography}
\title{Transforms a set of points by a homography.
xy_mat: a (n_points, 2) matrix of data to
homography: a (3, 3) matrix following the conventions of a 2d homography}
\usage{
transform_points_by_homography(xy_mat, homography)
}
\description{
Returns: the (n_points, 2) transformed points
}
