% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter-tuning.R
\name{visualize_metric_vs_1_parameter}
\alias{visualize_metric_vs_1_parameter}
\title{Visualize one metric versus one or more paramter values}
\usage{
visualize_metric_vs_1_parameter(
  results_df,
  parameter_df,
  attributes_to_visualize,
  eval_function,
  smoothing_bandwidth = 5,
  title = ""
)
}
\arguments{
\item{results_df}{The results of trials, one per row}

\item{attributes_to_visualize}{Which paramters to visualize the metrics versus}

\item{eval_function}{How to evalute the quality of the result}

\item{smoothing_bandwidth}{Width of the kernel used to smooth the trend line. May be a scalar or one value per attribute being visualized.}

\item{title}{Title of the plot}

\item{paramter_df}{The paramters used to generate the results, one row per trial}
}
\description{
Visualize one metric versus one or more paramter values
}
