% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata-extraction_imagery_perimage.R
\name{extract_altitude_asl}
\alias{extract_altitude_asl}
\title{Returns altitude above sea level (asl) in meters}
\usage{
extract_altitude_asl(
  exif,
  candidate_asl_columns = c("AbsoluteAltitude", "GPSAltitude")
)
}
\arguments{
\item{exif}{the exif metadata file}
}
\value{
altitude above sea level in meters
}
\description{
Returns altitude above sea level (asl) in meters to include in image-level metadata collation.
}
\examples{
extract_altitude_asl(exif)

}
