% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-imagery-metadata_emp.R
\name{extract_aperture}
\alias{extract_aperture}
\title{Returns aperture (format: xxxxx)}
\usage{
extract_aperture(exif)
}
\arguments{
\item{exif}{the exif metadata file}
}
\value{
aperture (format: xxxxx)
}
\description{
Returns aperture to include in image-level metadata collation.
}
\examples{
extract_aperture(exif)

}
