% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-imagery-metadata_emp.R
\name{extract_exposure}
\alias{extract_exposure}
\title{Returns exposure (units: seconds)}
\usage{
extract_exposure(exif)
}
\arguments{
\item{exposure}{exposure time in seconds}
}
\value{
exposure (units: seconds)
}
\description{
Returns exposure time in seconds to include in image-level metadata collation.
}
\examples{
extract_exposure(exif)

}
