% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata-extraction_imagery_dataset.R
\name{extract_exposure_summary}
\alias{extract_exposure_summary}
\title{Extracts the median exposure time and its coefficient of variation}
\usage{
extract_exposure_summary(metadata)
}
\arguments{
\item{metadata}{the metadata metadata file}
}
\value{
the median exposure time across all images in the dataset. Units: sec
}
\description{
The median exposure time across all images in the dataset. Units: sec
}
\examples{
extract_exposure(metadata)

}
