% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata-extraction_imagery_dataset.R
\name{extract_file_size}
\alias{extract_file_size}
\title{Extracts total file size}
\usage{
extract_file_size(exif)
}
\arguments{
\item{exif}{the exif metadata file}
}
\value{
total file size
}
\description{
Total file size of all images in set. Units: GB.
}
\examples{
extract_file_size(exif)

}
