% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-dataset-metadata_emp.R
\name{extract_image_count}
\alias{extract_image_count}
\title{Extract number of images in an exif}
\usage{
extract_image_count(exif)
}
\arguments{
\item{exif}{the exif metadata file}
}
\value{
image count
}
\description{
Pulls the image count to include in dataset-level metadata
}
\examples{
extract_image_count(exif)

}
