% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-imagery-metadata_emp.R
\name{extract_iso}
\alias{extract_iso}
\title{Returns ISO}
\usage{
extract_iso(exif)
}
\arguments{
\item{exif}{the exif metadata file}

\item{iso}{ISO}
}
\value{
ISO
}
\description{
Returns ISO to include in image-level metadata collation.
}
\examples{
extract_iso(exif)

}
