% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-imagery-metadata_emp.R
\name{extract_lat_lon}
\alias{extract_lat_lon}
\title{Extract latitude and longitude of image collection location}
\usage{
extract_lat_lon(exif)
}
\arguments{
\item{exif}{the exif metadata file}

\item{exif_coordinates}{creates data.frame with coordinates for each image}

\item{lat}{latitude of image collection}

\item{lon}{longitude of image collection}

\item{lat_lon}{data.frame of latitude and longitude of image collection}
}
\value{
latitude and longitude of image collection location
}
\description{
Pulls the latitude and longitude of image collection location (Format: dd.dddddddd) to include in image-level metadata collation
}
\examples{
extract_lat_lon(exif)

}
