% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-imagery-metadata_emp.R
\name{extract_metadata_emp}
\alias{extract_metadata_emp}
\title{Returns a collection of image-level metadata parameters}
\usage{
extract_metadata_emp(exif_filepath)
}
\arguments{
\item{exif}{the exif file being processed}

\item{dataset_id}{dataset id}

\item{datatime_local}{local date and time of image collection}

\item{lat_lon}{a data.frame of latitude and longitude of image collection}

\item{rtk_fix}{if a column for "RtkFlag" is in the exif data, rtk_fix is true if the column value is 50 and false for any other value. If there isn't an "RtkFlag" column, rtk_fix is false.}

\item{accuracy}{a data.frame of the accuracy_x (longitude accuracy standard deviation in meters), accuracy_x (latitude standard deviation in meters), and accuracy_z (height standard deviation in meters) values for each image}

\item{pitch_roll_yaw}{a data.frame of the camera_pitch (degrees up from nadir), camera_roll (degrees clockwise from up), and camera_yaw (degrees right from true north) values for each image}

\item{exposure}{exposure time in seconds}

\item{aperture}{aperture (format: xxxxx)}

\item{iso}{ISO}

\item{white_balance}{either auto or manual white balance}

\item{received_image_path}{the received image path}

\item{altitude_asl}{altitude above sea level in meters}

\item{metadata}{a data.frame of dataset_id, datatime_local, lat, lon, rtk_fix, accuracy_x, accuracy_y, accuracy_z, camera_pitch, camera_roll, camera_yaw, exposure, aperture, iso, white_balance, received_image_path, and altitude_asl}
}
\value{
a data.frame of dataset_id, datatime_local, lat, lon, rtk_fix, accuracy_x, accuracy_y, accuracy_z, camera_pitch, camera_roll, camera_yaw, exposure, aperture, iso, white_balance, received_image_path, and altitude_asl
}
\description{
Returns altitude above sea level (asl) in meters to include in image-level metadata collation.
}
\examples{
extract_metadata_emp(exif)

}
