% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagery-metadata-extraction_emp.R
\name{extract_pitch_roll_yaw}
\alias{extract_pitch_roll_yaw}
\title{Returns camera pitch, camera roll, and camera yaw. Units: degrees.}
\usage{
extract_pitch_roll_yaw(exif)
}
\arguments{
\item{exif}{the exif metadata file}
}
\value{
camera pitch, camera roll, and camera yaw. Units: degrees.
}
\description{
Returns camera_pitch (Units: deg, degrees up from nadir), camera_roll (Units: deg, degrees clockwise from up), and camera_yaw (Units: deg, degrees right from true north) to include in image-level metadata collation.
}
\examples{
extract_pitch_roll_yaw(exif)

}
