% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-imagery-metadata_emp.R
\name{extract_rtk_fix}
\alias{extract_rtk_fix}
\title{Returns rtk_fix of image}
\usage{
extract_rtk_fix(exif)
}
\arguments{
\item{exif}{the exif metadata file}
}
\value{
rtk fix of image, true/false
}
\description{
Pulls the rtk fix to include in image-level metadata collation. Return TRUE if there is a RtkFlag value of 50, and FALSE otherwise, i.e. other RtkFlag value, or none. IMPORTANT NOTE!!!!!! THIS FUNCTION ASSUMES DJI DRONES WERE USED. OTHER DRONE MAKES WILL REQUIRE AN UPDATED FUNCTION.
}
\examples{
extract_rtk_fix(exif)

}
